Okay so this is quite funny. This initial server may have been slightly too simple.
If you look at these wireshark logs, you'll see that the HTTP response is
actully listed before the request...which is the opposite of what you
would normally expect.

So why is that? Well I did tell you this was a very very simple server.
It's actually a little bit too simple. What I'm doing is sending the HTTP
response as soon as the TCP connection is established. I'm not actually
waiting for the client to send me a request at all! So I am indeed sending
the response before receiving the request.

And as we saw - the website works in chrome, so it basically worked.

But it is definitely not how HTTP is supposed to work. And it does have
some issues. Obviously most of the time you need to look at the request
before you know how you want to respond. Right now I'm not doing that.
I just return the same exact HTTP response regardless. For example if you
look at the request for the favicon in thr chrome network tools...you'll
see the same exact HTML as we see in the page.

Moreover, an HTTP client can choose to re-use a TCP connection to make
more requests. But that won't work with our current setup, because I only
ever send the response when a new TCP connection is established.